refactor: use u32 for line numbers throughout
tracing::Metadata::line() returns Option<u32>. The i64 type was only
needed at the SQLite binding boundary. int_or_null is now generic over
Into<i64>, and opt_u32 handles the read path with try_from validation.

Also: fix missing Timestamp::now() call, fix two clippy warnings in
collector.rs.

+++ aperture-http/src/dto/log.rs
@@ -86,7 +86,7 @@ pub struct LogEventResponse {
     /// Source file, if available.
     pub file: Option<String>,
     /// Source line, if available.
-    pub line: Option<i64>,
+    pub line: Option<u32>,
     /// All structured fields as a JSON object, if any.
     pub fields: Option<serde_json::Value>,
 }
@@ -127,7 +127,7 @@ pub struct LogSpanResponse {
     /// Source file, if available.
     pub file: Option<String>,
     /// Source line, if available.
-    pub line: Option<i64>,
+    pub line: Option<u32>,
     /// When the span started.
     pub started_at: Timestamp,
     /// When the span ended, if it did.
